Cast receiver if necessary when rewriting trait calls to impl method
The self parameter type may be incompatible with the trait type.

  trait T { self: S => def foo = 1 }

The $self parameter type of T$class.foo is S, which may be unrelated
to T. If we re-write a call to T.foo to T$class.foo, we need to cast
the receiver to S, otherwise we get a VerifyError.