SI-9413 Type check default argument expressions only once
Default argument expressions were type-checked twcie: once in the
ValDef of the original method, and once in the default getter method.
The context used for typing the parameter ValDef was wrong, causing
the issue described in SI-9413.

There's no point in type-checking the parameter ValDef's rhs, so we
just replace it by EmptyTree.