Fix expected type when type checking argument of value class
Example

    class A(val x: String) extends AnyVal
    class B(val x: A)      extends AnyVal

    class C {
      def t = {
        val a = new A("bcc")
        val b = new B(a)
        b.x
      }
    }

During erasure, when (re-)type-checking `new B(a)` in `typed1` of
`transform/Erasure`, we get the a tree `InjectDerivedValue` representing
the `new B(a)`.

Before this patch, the expected type for type checking `a` was the
erased type of B's value field, which is String (erasure replaces type
`A` by `String`). The tree `a` has type `ErasedValueType(A, String)`.
During erasure's `adapt` phase, the expected type did not match. Since
the tree's type is an (erased) value type, the adapt method would
introduce a box creation (i.e., `new A(a)`). This is clearly wrong.

The solution here is that the expected type for type-checking the `a`
tree should be `ErasedValueType(A, String)`. This is actually the type
of B's value at this phase, so we just need to make sure not to erase
that down to A. This is what specialErasure does.

Bytecode for `t` after this patch (if you also enable `-opt:l:project`,
otherwise there's some redundant local variables):

    public t()Ljava/lang/String;
     L0
      LINENUMBER 6 L0
      LDC "bcc"
     L1
      LINENUMBER 8 L1
      ARETURN