Desugar local lazy vals into a box with a volatile boolean init field
Inspired by dotc, desugar a local `lazy val x = rhs` into

    val x$lzy = new scala.runtime.LazyInt()
    def x(): Int = {
      if (!x$lzy.initialized) {
        x$lzy.synchronized {
          if (!x$lzy.initialized) {
            x$lzy.initialized = true
            x$lzy.value = rhs
          }
        }
      }
      x$lzy.value
    }

Note that the 2.11 decoding (into a local variable and a bitmap) also
creates boxes for local lazy vals, in fact two for each lazy val:

    def f = {
      lazy val x = 0
      x
    }

desugars to

    public int f() {
        IntRef x$lzy = IntRef.zero();
        VolatileByteRef bitmap$0 = VolatileByteRef.create((byte)0);
        return this.x$1(x$lzy, bitmap$0);
    }

    private final int x$lzycompute$1(IntRef x$lzy$1, VolatileByteRef bitmap$0$1) {
        C c = this;
        synchronized (c) {
            if ((byte)(bitmap$0$1.elem & 1) == 0) {
                x$lzy$1.elem = 0;
                bitmap$0$1.elem = (byte)(bitmap$0$1.elem | 1);
            }
            return x$lzy$1.elem;
        }
    }

    private final int x$1(IntRef x$lzy$1, VolatileByteRef bitmap$0$1) {
        return (byte)(bitmap$0$1.elem & 1) == 0 ? this.x$lzycompute$1(x$lzy$1, bitmap$0$1) : x$lzy$1.elem;
    }

An additionl problem with the above encoding is that the `lzycompute`
method synchronizes on `this`. In connection with the new lambda
encoding that no longer generates anonymous classes, captured lazy vals
no longer synchronize on the lambda object.

The new encoding solves this problem by synchronizing on the lazy
holder.

Resolves https://github.com/scala/scala-dev/issues/133