Use `foreachSymbolInType` for performance reasons
Instead of gathering all symbols in a set and using `foreach`, reuse a
foreach symbol traverser that will go through all the symbols of a type
and execute an operation on them.

This finishes up similar work done by @romanowski. Apparently, this
instance of `symbolsInType` was forgotten. As this function is not
useful anymore, it is removed for the posterity.