Improve performance of `Dependency`
`Dependency` was architectured in two phases. First, a traverser would
run in the tree of the compilation unit and populate all the found
dependencies. Second, the caller of the traverser would go through these
data structures and process them.

Since performance is a major feature of Zinc, this commit changes the
previous design and mixes the previously described steps into one. This
is possible because they don't depend on each other and processing
dependencies does not require any kind of ordering.

This is likely to have a significant impact in the performance of
`Dependency` memory-wise for the following reasons:

* At the moment of processing the class dependency, that object is
  already hot in the cache and its access is for free.
* We economise the amount of memory required for holding the populated
  temporary data structures. For larger compilation units, this data would
  be large and likely to be promoted from young generations to older
  generations, thus consuming unnecessary GC time.

Note that `addClassDependency` and `addTopLevelImportDependency` cache
the added dependency unlike the previous version, which would just add
to the hash set even though the key was already there.

This rewrite of `Dependency` tries to touch as less lines as possible to
preserve the git history for future archeologists of this codebase. It
also adds a lot of new comments explaining the underlying algorithm.

@romanowski has performed a similar job before when he rewrote the
temporary data structures to use `Iterator`s instead of the underlying
sets. I'm unsure of the real impact of this change in comparison with
his, but it should at least be a little bit faster.