Raise the timer of names performance
Commit 929b7589bbf02140b17de04106b7162eebc5db21 introduced a decrease of
20 seconds -- from 30 seconds to 10 seconds -- because it was being
optimistic.

This timer is the responsible of making this test pass, and though 10
seconds is a very good time and we've passed it, it can happen that
compilation lasts more than expected and fails the tests.

For that reason, and to find a good tradeoff, this commit raises the
timer from 10 seconds to 20 seconds, 1/3 less of what we had originally
before this PR.