Initialize statistics per global
As described in the bug report, statistics are currently shared by
different instances of `Global` because they are located in objects.

This approach has an important disadvantage.  Statistics from different
global instances step on each other's toes, rendering the statistics
infrastructure unusable in most of the Scala developers' workflows
(think sbt running in parallel several compilations). When the data is
not useless, it produces races conditions and kills compilation.

This patch addresses the issue by creating a protected abstract
`statistics` field in `SymbolTable` that is overriden in `Global` and
`JavaUniverse` and hence available to most of the compiler internals.
This object contains all the statistics at the known definition site
(`Global` defines more statistics that are created in `scala.tools.nsc`
instead of `scala.reflect`). All statistics are accessible via `import
statistics._`, following the existing idiom for other pieces of the
compiler like `definitions`.

The main goal of this implementation was to avoid moving the definition
of statistics from the places where they are used. Currently, they are
defined alongside the classes that have the call-sites (see, for
example, `Implicits.scala`, `Macros.scala` or `Types.scala`).

This commit produces a binary incompatible change because of the removal
of `IOStats`. This change is whitelisted because `IOStats` is `private[io]`.

`AliasingFrame` and `IOStats` are removed because the stats defined
there are completely disconnected from `Global` and `SymbolTable`. Later
approaches will try to bring `IOStats` back to life again, since it's
the most useful of the two.

Fixes scala/bug#10460.