Don't add original attachment for constant folded trees
Scalac seems to run twice constant-folding in certain scenarios.
Consider the following example:

```scala
object A { final val x = 1 }

object B {
  def main(args: Array[String]) = assert(args(0).toInt == A.x )
}
```

The tree `A.x` in `main` is constant folded because `x` is a constant,
but for some reason the resulting tree `Literal(Constant(1))` get
constant-folded again.

The previous logic would add the original tree attachment on the
constant-folded tree, and would destroy the previous annotation that
gets copied by `treeCopy.Literal`.

Therefore, to avoid this issue, we don't update the original attachment
if there already exists one.

In the future, it would be worthwhile to figure out why scalac tries to
constant-fold twice the same tree.