Properly report problems to sbt
Dotty has its own logic for displaying problems with the proper file
path, position, and caret, but if we store this information in
Problem#message we end up with duplicated information in the output
since Zinc will prepend/append similar things (see
sbt.internal.inc.ProblemStringFormats). So far, we worked around this in
Dotty by using an empty position in the sbt bridge reporter, but this
means that crucial semantic information that could be used by a Build
Server Protocol implementation and other tools is lost.

Thanks to https://github.com/sbt/zinc/pull/588 we can now fully
customize how the message is displayed to the user using
Problem#rendered, so we can now store the position information without
any adverse effect.