Collapse positions in macro expansions after duplication
Macro expansion changes positions in the expanded tree to offset
(in `validateResultingTree`).

For spliced, non-duplicated trees this can modify the positions of
the expandee, which means that the `MacroExpansionAttachment` attachment
doesn't contain the original tree with range positions.

This PR collapses positions of macro expansions after they are
duplicated. The duplication is already in place to avoid structural
sharing in macro-generated trees, a "mistake" that's easy to do by
macro authors.